Move SecondWind to trash
This only works if we can make it super obvious to the player that
they're in the "need revival" mode. Would be nice if we could make
client go into third person.

Only way I see this working is if we actually simulate this ourselves -
e.g. tp'ing player in spectator mode momentarily, showing them an npc
clone of themselves, and then back to first person... work I'm
definitely not going to do lol