save and restore exp separately
So we can show the player's level while in the minigames spawn (since
experience level is "global")

We only need to save when player enters minigames spawn (a player should
not be able to join minigames while in a survival world). We restore
whenever they return to minigames spawn and when they teleport back to a
survival world (in case they directly teleport to someone in survival).